Added some selection, moving, and linking commands from the user, as well as created an output to the user in text in the top left corner.
Hold Shift to add a box, press L to enter link mode, and press M to enter move mode. Any key press of 1-9 will select a box, turning it white.

To link two nodes, press L to enter link mode, press the index of the first to link, then the second to link.
To move a node, press M, select the node by pressing its index, and then tap with the LButton where you want it to go.